requirements.txt
Babel==2.1.1
Django==1.8.4
django-appconf==1.0.1
django-geoposition==0.2.2
django-localflavor==1.2
django-localflavor-br==1.0
django-social-auth==0.7.28
django-social-widgets==0.3.0
django-tinymce==2.0.6
django-tinymce-images==0.1.2=
html5lib==0.9999999
httplib2==0.9.2
i18n==0.2
i18nize-templates==0.1.dev20131011
oauth2==1.9.0.post1
Pillow==2.9.0
pip==7.1.2
py==1.4.30
python-openid==2.2.5
pytz==2015.7
setuptools==18.2
six==1.10.0
wheel==0.24.0